Checkin for Bolt 1.2 GA release. New features added: -Function parity between AMP and OpenCL Backends. -Transform iterator support in OpenCL backend. -Permutation iterator support in AMP backend. -Performance optimizations.

+++ examples/DeviceVector/deviceVector.cpp
@@ -1,5 +1,5 @@
 /***************************************************************************                                                                                     
-*   Copyright 2012 - 2013 Advanced Micro Devices, Inc.                                     
+*   © 2012,2014 Advanced Micro Devices, Inc. All rights reserved.                                     
 *                                                                                    
 *   Licensed under the Apache License, Version 2.0 (the "License");   
 *   you may not use this file except in compliance with the License.                 
@@ -52,7 +52,7 @@ int _tmain( int argc, _TCHAR* argv[ ] )
     pdV.reset( );
 
     std::cout << "Device Vector contents: " << std::endl;
-    for( size_t i = 0; i < vecSize; ++i )
+    for( int i = 0; i < vecSize; ++i )
     {
         std::cout << dV[ i ] << ", ";
     }
